Split tests into saas and onprem
