+++ test/integration/pytest_slc_test_code.py.txt
@@ -1,0 +1,37 @@
+import pyexasol
+import pytest
+from exasol.python_extension_common.deployment.language_container_validator import temp_schema
+from exasol.python_extension_common.deployment.language_container_builder import (
+    LanguageContainerBuilder, find_path_backwards)
+
+@pytest.fixture(scope='session')
+def slc_builder() -> LanguageContainerBuilder:
+    with LanguageContainerBuilder('test_container') as container_builder:
+        project_directory = find_path_backwards("pyproject.toml", "{project_file}").parent
+        container_builder.prepare_flavor(project_directory)
+        yield container_builder
+
+@pytest.fixture(scope='session')
+def language_alias() -> str:
+    return "{main_language_alias}"
+
+def assert_udf_running(conn: pyexasol.ExaConnection, lang_alias: str) -> None:
+    with temp_schema(conn) as schema:
+        udf_name = 'TEST_UDF'
+        udf_create_sql = (
+            f'CREATE OR REPLACE {{lang_alias}} SCALAR SCRIPT "{{schema}}"."{{udf_name}}"() '
+            'RETURNS BOOLEAN AS '
+            'def run(ctx): '
+            'return True '
+            '\n/'
+        )
+        conn.execute(udf_create_sql)
+        result = conn.execute(f'SELECT "{{schema}}"."{{udf_name}}"()').fetchall()
+        assert result[0][0] is True
+
+def test_deploy_slc(deploy_slc, deployed_slc, backend_aware_database_params):
+    # We will activate the SLC also with an alternative alias and check that both
+    # the main and the alternative alias work.
+    deploy_slc("{alt_language_alias}")
+    for lang_alias in ["{main_language_alias}", "{alt_language_alias}"]:
+        assert_udf_running(pyexasol.connect(**backend_aware_database_params), lang_alias)
